Separate tracking map into its own header
